requirements.txt
tqdm
torch
click
pyyaml
numpy
torchvision
diffusers
Pillow
scipy
munch
transformers
torchmetrics
scikit-image
opencv-python
sentencepiece
protobuf
accelerate
gradio